Refactor drift path helpers to keep module under LOC budget
